Flipping the pancake
The pancake now flips! But when it is facing downwards on the pan it will try to go into the pan and will not move. There are now walls on either side of the scene so the pancake never flies of screen left of right, and when it falls off it respawns at 0,0,0. A torque spins the pancake randomly when it is flipped. I made the pancake smaller and the pan smaller to make catching it more challenging.  I had to remove the flipped boolean from changing to true in order to be able to flip the pancake more than once. This means that it can be flipped in mid air.