Sprites and new Mechanics
The game now has creative commons sprites. A picture of a kitchen as the background and clip art for the pan and pancake. To solve the problem with the pancakes being unable to flip correctly when upside down on the pan, clicking mouse2 will now cause the pancakes to flip right side up. The pan now also moves on the y axis with the mouse. I'm not sure how I feel about this change, test it and see what you think.