Added background to other scenes and tweaked code.
Took the kitchen background, darkened it, and added that darkened copy to the main menu and how to scenes. I also tweaked the code so that the pancake is righted when you hold down the left mouse button, plus I locked the pan to the y-axis to prevent weird interactions with the pan's and pancake's box collider.